Add files via upload
Restrictive detail codes are codes that are added to appointment slots that prevent patients from booking into them. They are essentially "blocked" and make it so that only the provider or clinic staff can book the appointments instead of the central appointment line or online booking. The problem with this is that it severely decreases the access to care that patients need. This chart retroactively looked at the total number of appointments (total appointments = open + frozen + canceled + booked appointments) and filters by a slicer the appointments that were restricted via a detail code. It then looks at the percentage of appointments that were actually used by the office staff. This was reviewed on a monthly basis to see which clinics were making effective use of their detail codes as well as identify which clinics needed to tighten up their booking procedures.