Add files via upload
The appointment utilization file shows the utilization of appointments each month. Utilization is defined as the percentage of total appointments that were booked, expressed as a percentage. Along with the utilization of appointments, we can also see the appointment type mix (there are essentially two types of appointments--acute appointments and routine visit appointments). Combining this data along with the average days to care can help identify where we may need to adjust the mix of appointments.