Add files via upload
This shows the average days to care that a patient would wait to get seen for a given month. This is often looked at in tandem with the appointment utilization and appointment mix to determine if there needs to be more of one type of appointment offered instead of the other (for example, if there is a 60/40 split between routine and acute type appointments, with routine appointments being utilized at 25% and acute appointments being utilized at 87%, with average days to a routine appointment being 2 days and average day to an acute appointment being 4 days, we may need to shift that split to more of a 50/50 or even 40/60 split).