Add files via upload
A tool used to help plan weekly appointment offerings at a small outpatient clinic. This helps to determine the mix of appointments, the number of appointments offered each hour, what kinds of appointments offered each hour, percentages of face to face and virtual appointments, restricted appointments, and appointments available online. All of these metrics were used to "score" this facility which directly relates to the funding the facility receives.